// frontend/src/atoms/Button/Button.types.ts
import type { ButtonHTMLAttributes } from 'react'
export type ButtonVariant = 'primary' | 'danger' | 'ghost' | 'command'
export type ButtonSize    = 'sm' | 'md' | 'lg'

export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
  variant?:  ButtonVariant
  size?:     ButtonSize
  loading?:  boolean
  prefix?:   string
}